move all code to red/ubicall to begin to orgnize code behind generating plist `util.js`
