+++ red/api/index.js
@@ -23,7 +23,7 @@ var passport = require('passport');
 var ui = require("./ui");
 var nodes = require("./nodes");
 var flows = require("./flows");
-var plist = require("./plist");
+var plist = require("../ubicall/plist");
 var library = require("./library");
 var info = require("./info");
 var auth = require('./auth');
